code

+++ app/src/main/java/sk/sandeep/testapp/model/Current.kt
@@ -14,5 +14,5 @@ data class Current(
     val visibility: Int,
     val weather: List<Weather>,
     val wind_deg: Int,
-    val wind_speed: Int
+    val wind_speed: Double
 )
